Fully implemente the RESTful api. Added support for the Java-Proxy-Server as well so that maybe we could actually implement this thing before the poster-fair/demo-day comes upgit add --all

+++ UniTap-BackEnd/app/controllers/api/company_cards_controller.rb
@@ -2,22 +2,54 @@ class Api::CompanyCardsController < ApplicationController
   skip_before_filter :verify_authenticity_token
   
   def index
-  	render json: CompanyCard.all
+    if !(params[:cardhash].nil? || params[:phonecrc].nil?)
+  	  if !(CompanyCard.find_by(cardhash: params[:cardhash], phonecrc: params[:phonecrc]).nil?)
+        render status: 200, json: {
+          message: "Welcome to our business!"
+        }
+      else
+        render status: 422, json: {
+          message: "Request Denied."
+        }
+      end
+    elsif !(params[:id].nil?)
+      if(CompanyCard.exists?(params[:id]))
+        card = CompanyCard.find(params[:id])
+        render status: 200, json: {
+          company_cards: card
+        }.to_json
+      end
+    elsif !(params[:cardhash].nil?) || !(params[:phonecrc].nil?)
+      render status: 422, json:{
+        message: "Your request makes no sense."
+      }
+    else
+      render status: 200, json: CompanyCard.all
+
+    end
+
   end
   
   ##
   # display a particular card
   #
   def show
-  	if(CompanyCard.exists?(params[:id]))
-      card = CompanyCard.find(params[:id])
-      render status: 200, json: {
+  	if(!params[:id].nil?)
+      if(CompanyCard.exists?(params[:id]))
+        card = CompanyCard.find(params[:id])
+        render status: 200, json: {
         company_cards: card
       }.to_json
+      
+      else
+        render status: 422, json: {
+          message: "Unable to process your request of the server."
+        }.to_json
+      end
     else
-      render status: 422, json: {
-        message: "Could not find that card."
-      }.to_json
+        render status: 422, json: {
+          message: "Unable to process your request of the server."
+        }.to_json
     end
   end
 
@@ -25,32 +57,44 @@ class Api::CompanyCardsController < ApplicationController
   # Create a new card for the company's database
   #
   def create
-    card = CompanyCard.new(card_params)
-    if card.save
-      render status: 200, json: {
-        message: "Successfully created new Company-Card.",
-        company_cards: card
+    if (params[:cardhash].nil?) || (params[:phonecrc].nil?)
+      render status: 422, json: {
+        message: "Card format incorrect.",
       }.to_json
     else
-      render status: 422, json: {
-  	    message: "Unable to process your request of the server."
-      }.to_json
-  	end
+      card = CompanyCard.new(card_params)
+      if card.save
+        render status: 200, json: {
+          message: "Successfully created new Company-Card.",
+          company_cards: card
+        }.to_json
+      else
+        render status: 422, json: {
+       	  message: "Unable to process your request of the server."
+        }.to_json
+  	  end
+    end
   end
 
   ##
   # Delete a certain card in the database
   #
   def destroy
-  	if(CompanyCard.exists?(params[:id]))
-      card = CompanyCard.find(params[:id])
-      card.destroy
-      render status: 200, json:{
-        message: "Successfully deleted that card."
-      }.to_json
+    if !(params[:id].nil?)
+      if (CompanyCard.exists?(params[:id]))
+        card = CompanyCard.find(params[:id])
+        card.destroy
+        render status: 200, json:{
+          message: "Successfully deleted that card."
+        }.to_json
+      else
+        render status: 422, json:{
+          message: "Unable to process your request of the server."
+        }.to_json 
+      end
     else
       render status: 422, json:{
-        message: "Could not delete that card."
+        message: "Unable to process your request of the server."
       }.to_json 
     end
   end
@@ -59,17 +103,23 @@ class Api::CompanyCardsController < ApplicationController
   # Update a certain card in the company's database
   #
   def update
-    if(CompanyCard.exists?(params[:id]))
-      card = CompanyCard.find(params[:id])
-      card.update(card_params)
-      render status: 200, json:{
-        message: "Successfully updated",
-        company_cards: card
-      }.to_json
+    if !(params[:id].nil?)
+      if(CompanyCard.exists?(params[:id]))
+        card = CompanyCard.find(params[:id])
+        card.update(card_params)
+        render status: 200, json:{
+          message: "Successfully updated",
+          company_cards: card
+        }.to_json
+      else
+        render status: 422, json:{
+          message: "Unable to process your request of the server."
+        }.to_json
+      end
     else
       render status: 422, json:{
-        message: "Could not update that card"
-      }.to_json
+          message: "Unable to process your request of the server."
+        }.to_json
     end
   end
 
@@ -78,6 +128,6 @@ class Api::CompanyCardsController < ApplicationController
   #
   private
   def card_params
-    params.require("company_card").permit("userid")
+    params.permit(:company_card, :cardhash, :phonecrc)
   end
 end
